requirements.txt
decord==0.6.0
gensim==4.3.3
llava==1.2.2.post1
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.5
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.2.1
Requests==2.32.3
rouge_score==0.1.2
scipy==1.15.2
seaborn==0.13.2
sentence_transformers==3.4.1
spacy==3.8.4
torch==2.1.2
torch==2.5.1
torchvision==0.16.2
torchvision==0.20.1
tqdm==4.66.5
transformers==4.37.2
ultralytics==8.3.78
